pyproject.toml
[project]
name = "celeste-core"
version = "0.1.0"
description = "Shared core for Celeste projects: config, enums, base classes, types, and utils"
requires-python = ">=3.13"
dependencies = [
  "pydantic>=2.8.2",
  "pydantic-settings>=2.3.0",
  "python-dotenv>=1.1.1",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
